requirements.txt
pandas~=2.2
numpy~=1.26
google-cloud-storage~=2.18
fastapi~=0.115
uvicorn[standard]~=0.30
google-cloud-aiplatform~=1.38
google-auth~=2.25
